follow up to removing some consensus params
Refs #2382

+++ consensus/reactor_test.go
@@ -196,7 +196,7 @@ func newMockEvidencePool(val []byte) *mockEvidencePool {
 }
 
 // NOTE: maxBytes is ignored
-func (m *mockEvidencePool) PendingEvidence(maxBytes int) []types.Evidence {
+func (m *mockEvidencePool) PendingEvidence(maxBytes int64) []types.Evidence {
 	if m.height > 0 {
 		return m.ev
 	}
