+++ consensus/state.go
@@ -953,14 +953,16 @@ func (cs *ConsensusState) createProposalBlock() (block *types.Block, blockParts 
 	// bound evidence to 1/10th of the block
 	evidence := cs.evpool.PendingEvidence(types.MaxEvidenceBytesPerBlock(maxBytes))
 	// Mempool validated transactions
-	txs := cs.mempool.ReapMaxBytesMaxGas(types.MaxDataBytes(maxBytes, cs.state.Validators.Size(), len(evidence)), maxGas)
-
+	txs := cs.mempool.ReapMaxBytesMaxGas(types.MaxDataBytes(
+		maxBytes,
+		int64(cs.state.Validators.Size()),
+		int64(len(evidence)),
+	), maxGas)
 	proposerAddr := cs.privValidator.GetAddress()
 	block, parts := cs.state.MakeBlock(cs.Height, txs, commit, evidence, proposerAddr)
 
 	return block, parts
 }
-
 
 // Enter: `timeoutPropose` after entering Propose.
 // Enter: proposal block and POL is ready.
